Remove unused crowdmap class
